+++ NEWS.md
@@ -5,6 +5,7 @@
    active line across all cells. The edit mode button is removed. (#2691, #2682)
  * Command prompt is focused when its language is changed (#2678)
  * Welcome message is replaced when starting a new session (#2669)
+ * mini.js detects notebook and version in `notebook.R` URLs (#2650)
 
 ### Bugfixes
  * Forking a foreign notebook could fail with the latest rcloud-gist-services (#2679)